+++ src/components/pages/home/CryptoCards.jsx
@@ -1,17 +1,31 @@
-import React, { Suspense } from 'react'
+import React from 'react'
 import useFetch from '../../../lib/hooks/useFetch';
+import Loader from '../../shared/Loader';
 
 const CryptoCards = () => {
     const { data, error, loading } = useFetch('https://api.coindesk.com/v1/bpi/currentprice.json', 'CryptoCards Component');
     if (loading) {
-        return <div>Loading...</div>
+        return <div className='card-container'>
+            <Loader style={{
+                width: "200px",
+                height: "150px"
+            }} />
+            <Loader style={{
+                width: "200px",
+                height: "150px"
+            }} />
+            <Loader style={{
+                width: "200px",
+                height: "150px"
+            }} />
+        </div>
     }
 
     if (error) {
         return <div>{error}</div>
     }
     return (
-        <div className='card-container'> 
+        <div className='card-container'>
             {
                 data && data.bpi && Object.keys(data.bpi).map((key, index) => (
                     <div key={index} className="card">
