+++ src/components/pages/home/Home.jsx
@@ -4,7 +4,6 @@ import Graph from './Graph';
 import CryptoCards from './CryptoCards';
 
 const Home = () => {
-  const { data, error, loading } = useFetch('https://jsonplaceholder.typicode.com/posts', 'Home Component');
   return (
     <div className='content-page'>
       <div className="hero-header">
