+++ src/components/shared/style.css
@@ -1,0 +1,29 @@
+.loader-container {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #899489;
+  border-radius: 16px;
+}
+.loader {
+    width: 100px; 
+    height: 100px;
+    border-radius: 50%;
+    border: 5px solid transparent;
+    border-top-color: #4c983f; 
+    animation: spin 1s linear infinite;
+}
+
+  /* Shiny loading effect */
+  @keyframes spin {
+    from {
+      transform: rotate(0deg);
+      box-shadow: 0 0 0 rgba(183, 210, 227, 0.8); 
+    }
+    to {
+      transform: rotate(360deg);
+      box-shadow: 0 0 20px rgba(139, 171, 192, 0.8);
+    }
+  }
