+++ src/index.css
@@ -183,6 +183,7 @@ body {
   gap: 10px;
   padding: 2px;
   border-radius: 0 10px 10px 0;
+  position: fixed;
 }
 
 .hero-header {
@@ -232,6 +233,7 @@ body {
 
 .card-container {
   display: flex;
+  flex-wrap: wrap;
   width: 80%;
   margin: auto auto;
   margin-top: 3rem;
@@ -287,11 +289,54 @@ footer {
     z-index: 10;
   }
 
+  .graph {
+    width: 100%;
+  }
+
   .content-page {
-    margin-left: 0;
+    margin-left: 20px;
+  }
+
+  .hero-header h1 {
+    font-size: 1.5rem;
+  }
+
+  .graph .container {
+    min-width: 600px;
+  }
+
+  .hero-header button {
+    padding: 8px 16px;
+    font-size: 0.8rem;
+  }
+
+  .card-container {
+    width: 100%;
+  }
+
+  .card-container .card {
+    width: 30%;
   }
   
   .active {
     left: 0;
+  }
+}
+
+@media screen and (max-width: 480px) {
+  .card-container .card {
+    width: 45%;
+  }
+
+  .card-container h1 {
+    font-size: 0.8rem;
+  }
+
+  .card-container h2 {
+    font-size: 0.75rem;
+  }
+
+  .card-container .list {
+    font-size: 0.7rem;
   }
 }
