+++ src/lib/hooks/useWindowSize.js
@@ -1,0 +1,22 @@
+import { useEffect, useState } from "react";
+
+const useWindowSize = () => {
+    const [isMobile, setIsMobile] = useState(false);
+
+    useEffect(() => {
+        const handleResize = () => {
+            if (window.innerWidth <= 768) {
+                setIsMobile(true);
+            } else {
+                setIsMobile(false);
+            }
+        }
+        handleResize();
+        window.addEventListener('resize', handleResize);
+        return () => window.removeEventListener('resize', handleResize);
+    }, []);
+
+    return { isMobile };
+}
+
+export default useWindowSize;
